deleted lr shared component moved dummydata to test folder
